+++ apps/core/models/assignment.py
@@ -10,6 +10,7 @@ class Assignment(BaseModel):
         default_permissions = ()
 
     name = models.CharField(max_length=200)
+    database = models.CharField(max_length=100, null=True, blank=True)
 
     def __str__(self) -> str:
         return f"{self.name}"