+++ conf/entrypoint.sh
@@ -7,6 +7,7 @@ done
 
 python3 manage.py collectstatic --no-input
 python3 manage.py migrate
+python3 manage.py setup
 
 echo "$GITHUB_TOKEN" | docker login ghcr.io -u $GITHUB_USER --password-stdin